\name{pprRFA-package}
\alias{pprRFA-package}
\alias{pprRFA}
\docType{package}
\title{
\packageTitle{pprRFA}
}
\description{
\packageDescription{pprRFA}
}

\details{

The DESCRIPTION file:
\packageDESCRIPTION{pprRFA}
\packageIndices{pprRFA}
The package 
}
\author{
\packageAuthor{pprRFA}

Maintainer: \packageMaintainer{pprRFA}
}
\references{
}

\keyword{ package }

\seealso{
\link{zppr}, \link{ppr2gam} 
}
\examples{

}
